+++ Changes.md
@@ -9,6 +9,7 @@ Not yet released: v0.03-beta
 - Fixed an issue with WebPerl::JSObject::toperl()
   where JS objects were not being converted properly.
 - Added AJAX demo
+- Added WebPerl autoloading for script tags
 - ...